requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
openai>=1.0
Pillow>=10.0
PyYAML>=6.0
PyQt6>=6.6
pynput>=1.7
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
